feat: added blog post list to home route
