+++ prebuild.js
@@ -1,0 +1,10 @@
+/*
+
+# Pre-build tasks done in this script
+
+1. Clear the contents of src/routes/posts
+1. Plug all blog entries from blog_pages/ directory into src/routes/pages
+
+*/
+
+
