+++ src/routes/+page.svelte
@@ -1,5 +1,18 @@
+
+<script>
+import PostList from "../components/PostList.svelte";
+</script>
+
 # Ayden Diel
 
 Hi! I'm a developer from _Massachusetts_, and currently a **Computer Science** undergrad student at **UMass Boston**. Previously, I was a **SWE intern** at <span class="tag fidelity">Fidelity Investments</span>.
 
 I'm also the president of the <span class="tag umb">UMass Boston Computer Science Club</span>! [Check us out on LinkedIn](https://www.linkedin.com/company/umass-boston-computer-science-club/)
+
+## My Blog
+
+To see what I'm up to, check out my blog posts!
+
+![cookie monster pointing down](images/cookie-monster.gif)
+
+<PostList />
