+++ svelte.config.js
@@ -7,7 +7,7 @@ const config = {
   // Consult https://kit.svelte.dev/docs/integrations#preprocessors
   // for more information about preprocessors
   preprocess: [vitePreprocess(), mdsvex({
-    extensions: ['.md']
+    extensions: ['.md', '.svelte']
   })],
   extensions: ['.svelte', '.md'],
   kit: {
